separated search bar and filter components, made the mobile filter ui
